feat: Writing iterator to help debug isses with my max(total_minutes) being more than the time there is an a single week.
This feature finds which account it is specifically that has such a high count and allows me to see what the reason is.

fix: In my within_one_week function written earlier, i found that although specified to only pull accounts with enrollment times of more then 7 days. I did not account for users that have signed up and cancelled and then signed back up again multiple times. so I added and timedelta.days >=0 to fix this issue